feat: Add sketch for processing
Using touch designer is a PITA for this.
MinMax and mapping seems overly complex and I failed to understand
the UI/model how to properly read the value from the serial and
add isolate it to drive the phongs emit color. Very annoying
I tried to ask AI but the descriptions where also annoying
sincve the information it seems to have on the UI are outdated
Anyways moved back to the simplicity of good old processing.
I still used AI since I did not remember that the Serial needs
readStringUntil("\n") but these few lines of code are way better to
understand than the frustrating touchdesinger.